Fix cross-host quorum tracking.
The read and write replicators would improperly confirm with remote hosts that
their quorums matched.  They would only confirm the first host in the ring.

This commit fixes this behavior so that the replicator confirms that each remote
host's mapping of itself matches the mapping seen for that host by the
replicator.